+++ soapy_discovery.c
@@ -263,7 +263,7 @@ void soapy_discovery() {
 fprintf(stderr,"soapy_discovery\n");
   rtlsdr_count=0;
   SoapySDRKwargs *results = SoapySDRDevice_enumerate(NULL, &length);
-fprintf(stderr,"soapy_discovery: length=%d\n",length);
+fprintf(stderr,"soapy_discovery: length=%d\n",(int)length);
   for (i = 0; i < length; i++) {
     for (size_t j = 0; j < results[i].size; j++) {
       if(strcmp(results[i].keys[j],"driver")==0 && strcmp(results[i].vals[j],"audio")!=0) {
